requirements.txt
streamlit
audio-recorder-streamlit
faster-whisper
transformers
torch
librosa
spotipy
python-dotenv
numpy
scipy
soundfile
